Fixed page_link when creating new MenuItem
Cannot create menu item for page_link. Got javascript error:
`
An invalid form control with name='' is not focusable. <input type="text" class="form-control" placeholder="Internal slug. Ex: 'admin/page' (no quotes) for ':url'" for="link" required="">
`
Explanation: The browser tries to validate the input because it's required, even though it's required. But the input is not focusable, that's why it throws the error.

This piece of change solves that problem